refactor duplicate code home & search screen
